template.info: Add the new REQUIRES variable.
This variable will contain a space-separated list of all required
build-time and run-time dependencies, each named exactly as listed
on the SlackBuilds.org site.  It should not contain anything that
is part of Slackware itself.

If user interaction is mandatory (creation of user/group, etcetera)
or even recommended (ffmpeg, vlc, etcetera), REQUIRES should also
contain "%README%".  Automated tools should bail out and inform the
user that manual interaction is required when %README% is found.

The order of REQUIRES content is irrelevant.  Sorting (if desired)
is an implementation detail left to users and/or third-party tools.